Update frontend build assets (#631)
* Update gitignore to add vendor folder and ignore compiled js files

* Update and move package and gulp files to the project root folder

* cleanup file

* Update documentation

* update changelog
