+++ .gitignore
@@ -51,7 +51,10 @@ composer.lock
 /ecms_base/themes/custom/ecms/assets/data/data.json
 !/ecms_base/themes/custom/ecms/gulpfile.js
 !/ecms_base/themes/custom/ecms/components/**/*.js
+!/ecms_base/themes/custom/ecms/assets/vendor/**/*.js
 !/ecms_base/themes/custom/ecms/assets/scripts/**/*.js
+/ecms_base/themes/custom/ecms/assets/scripts/scripts-compiled-min.js
+/ecms_base/themes/custom/ecms/assets/scripts/scripts-compiled.js
 
 # ddev development site
 /develop
