+++ CHANGELOG.md
@@ -11,8 +11,10 @@ modified Semantic Versioning scheme. See the "Versioning scheme" section of the
 ## [Unreleased]
 ### Added
 - RIGA-670: Added D11 Patch: Migrate Process Trim
+- RIGA-652: Add vendor folder of package libraries
 
 ### Changed
+- RIGA-652: Update and move package and gulp files
 
 ### Deprecated
