+++ ecms_base/themes/custom/ecms/assets/patterns/00-atoms/00-base/__breakpoints.scss
@@ -45,7 +45,3 @@ $bp__text-collection--3col: $bp--large;
     @content;
   }
 }
-
-.poo-hole-one {
-  color: #ff00fs;
-}
