--- ecms_base/themes/custom/ecms/package.json
@@ -1,60 +1,0 @@
-{
-  "name": "ecms_patternlab",
-  "version": "1.0.1",
-  "description": "Pattern Lab node with Twig PHP Engine",
-  "main": "gulpfile.js",
-  "dependencies": {
-    "@babel/core": "7.12.9",
-    "@babel/preset-env": "7.12.7",
-    "@pattern-lab/cli": "5.13.1",
-    "@pattern-lab/core": "5.13.1",
-    "@pattern-lab/engine-twig-php": "5.13.1",
-    "@pattern-lab/uikit-workshop": "5.13.1",
-    "gulp": "4.0.2",
-    "gulp-autoprefixer": "^5.0.0",
-    "gulp-babel": "8.0.0",
-    "gulp-concat": "^2.6.1",
-    "gulp-eslint": "^4.0.0",
-    "gulp-hjson": "2.4.3",
-    "gulp-json-editor": "2.5.4",
-    "gulp-minify": "3.1.0",
-    "gulp-rename": "^1.2.3",
-    "gulp-sass": "^4.0.0",
-    "gulp-sass-glob": "^1.0.9",
-    "gulp-sass-import-json": "1.0.2",
-    "gulp-streamify": "1.0.2",
-    "gulp-stylelint": "^8.0.0",
-    "minimist": "1.2.5",
-    "prettier": "^1.14.2",
-    "stylelint": "^9.0.0",
-    "stylelint-config-sass-guidelines": "^5.0.0",
-    "stylelint-scss": "^3.0.0"
-  },
-  "babel": {
-    "presets": [
-      "@babel/preset-env"
-    ]
-  },
-  "keywords": [
-    "Pattern Lab",
-    "Atomic Web Design",
-    "Node",
-    "Gulp",
-    "Javascript"
-  ],
-  "repository": "https://github.com/pattern-lab/patternlab-node/tree/master/packages/edition-node-gulp",
-  "bugs": "https://github.com/pattern-lab/patternlab-node/issues",
-  "author": "Brian Muenzenmeyer",
-  "scripts": {
-    "gulp": "gulp -- ",
-    "patternlab": "patternlab"
-  },
-  "license": "MIT",
-  "engines": {
-    "node": ">=12.12.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "gitHead": "80f62be442223e09bafb30d0529cbd768e03f2ac"
-}
