+++ package.json
@@ -1,35 +1,47 @@
 {
   "name": "drupal-scaffold",
   "version": "1.0.0",
-  "description": "A Drupal 8 scaffold project",
+  "description": "A Drupal scaffold project using SDC's on top of some old PatternLab color theme logic.",
+  "main": "gulpfile.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/oomphinc/drupal-scaffold.git"
   },
   "license": "MIT",
   "dependencies": {
-    "babel-core": "^6.26.3",
-    "babel-preset-env": "^1.7.0",
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-config-prettier": "^3.0.1",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.0.3",
-    "eslint-plugin-prettier": "^2.6.2",
-    "eslint-plugin-react": "^7.11.1",
-    "gulp": "^3.9.1",
+    "@babel/core": "7.12.9",
+    "@babel/preset-env": "7.12.7",
+    "@pattern-lab/cli": "5.13.1",
+    "@pattern-lab/core": "5.13.1",
+    "@pattern-lab/engine-twig-php": "5.13.1",
+    "@pattern-lab/uikit-workshop": "5.13.1",
+    "gulp": "4.0.2",
     "gulp-autoprefixer": "^5.0.0",
-    "gulp-babel": "^7.0.1",
-    "gulp-eslint": "^4.0.2",
+    "gulp-babel": "8.0.0",
+    "gulp-concat": "^2.6.1",
+    "gulp-eslint": "^4.0.0",
+    "gulp-hjson": "2.4.3",
+    "gulp-json-editor": "2.5.4",
+    "gulp-minify": "3.1.0",
     "gulp-rename": "^1.2.3",
-    "gulp-sass": "^4.0.1",
+    "gulp-sass": "^4.0.0",
     "gulp-sass-glob": "^1.0.9",
-    "gulp-stylelint": "^7.0.0",
-    "node-sass": "^4.9.3",
+    "gulp-sass-import-json": "1.0.2",
+    "gulp-streamify": "1.0.2",
+    "gulp-stylelint": "^8.0.0",
+    "minimist": "1.2.5",
     "prettier": "^1.14.2",
-    "stylelint": "^9.4.0",
+    "stylelint": "^9.0.0",
     "stylelint-config-sass-guidelines": "^5.0.0",
-    "stylelint-scss": "^3.3.0",
-    "yargs": "^15.4.0"
+    "stylelint-scss": "^3.0.0"
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-env"
+    ]
+  },
+  "scripts": {
+    "gulp": "gulp -- ",
+    "patternlab": "patternlab"
   }
 }
